+++ mech_vmware_dvs/security_group_utils.py
@@ -1,0 +1,185 @@
+# Copyright 2015 Mirantis, Inc.
+# All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License"); you may
+# not use this file except in compliance with the License. You may obtain
+# a copy of the License at
+#
+#         http://www.apache.org/licenses/LICENSE-2.0
+#
+#    Unless required by applicable law or agreed to in writing, software
+#    distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+#    WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+#    License for the specific language governing permissions and limitations
+#    under the License.
+import abc
+
+from oslo_log import log
+import six
+
+from neutron.i18n import _LI
+
+LOG = log.getLogger(__name__)
+
+# protocol number according to RFC 1700
+PROTOCOL = {'icmp': 1,
+            'tcp': 6,
+            'udp': 17}
+
+
+@six.add_metaclass(abc.ABCMeta)
+class TrafficRuleBuilder(object):
+    action = 'ns0:DvsAcceptNetworkRuleAction'
+    direction = 'both'
+    reverse_class = None
+    _backward_port_range = (None, None)
+    _port_range = (None, None)
+
+    def __init__(self, spec_factory, ethertype, protocol, name=None):
+        self.factory = spec_factory
+
+        self.rule = self.factory.create('ns0:DvsTrafficRule')
+        self.rule.action = self.factory.create(self.action)
+
+        self.ip_qualifier = self.factory.create(
+            'ns0:DvsIpNetworkRuleQualifier'
+        )
+        self.ethertype = ethertype
+        if ethertype:
+            any_ip = '0.0.0.0/0' if ethertype == 'IPv4' else '::/0'
+            self.ip_qualifier.sourceAddress = self._cidr_spec(any_ip)
+            self.ip_qualifier.destinationAddress = self._cidr_spec(any_ip)
+
+        self.protocol = protocol
+        if protocol:
+            int_exp = self.factory.create('ns0:IntExpression')
+            int_exp.value = PROTOCOL.get(protocol, protocol)
+            int_exp.negate = 'false'
+            self.ip_qualifier.protocol = int_exp
+
+        self.name = name
+
+    def reverse(self):
+        """Returns reversed rule"""
+        name = 'reversed' + ' ' + (self.name or '')
+        rule = self.reverse_class(self.factory, self.ethertype,
+                                  self.protocol, name=name.strip())
+        rule.cidr = self.cidr
+        rule.port_range = self.backward_port_range
+        rule.backward_port_range = self.port_range
+        return rule
+
+    def build(self, sequence):
+        self.rule.qualifier = [self.ip_qualifier]
+        self.rule.direction = self.direction
+        self.rule.sequence = sequence
+        self.name = str(sequence) + '. ' + (self.name or '')
+        self.name = self.name.strip()
+        self.rule.description = self.name.strip()
+        return self.rule
+
+    @property
+    def port_range(self):
+        return self._port_range
+
+    @property
+    def backward_port_range(self):
+        return self._backward_port_range
+
+    @property
+    def cidr(self):
+        return self._cidr
+
+    def _port_range_spec(self, begin, end):
+        if begin == end:
+            result = self.factory.create('ns0:DvsSingleIpPort')
+            result.portNumber = begin
+        else:
+            result = self.factory.create('ns0:DvsIpPortRange')
+            result.startPortNumber = begin
+            result.endPortNumber = end
+        return result
+
+    def _cidr_spec(self, cidr):
+        try:
+            ip, mask = cidr.split('/')
+        except ValueError:
+            result = self.factory.create('ns0:SingleIp')
+            result.address = cidr
+        else:
+            result = self.factory.create('ns0:IpRange')
+            result.addressPrefix = ip
+            result.prefixLength = mask
+        return result
+
+    def _has_port(self, min_port):
+        if min_port:
+            if self.protocol == 'icmp':
+                LOG.info(_LI('Vmware dvs driver does not support '
+                             '"type" and "code" for ICMP protocol.'))
+                return False
+            else:
+                return True
+        else:
+            return False
+
+
+class IngressRule(TrafficRuleBuilder):
+    direction = 'incomingPackets'
+
+    def __init__(self, spec_factory, ethertype, protocol, name=None):
+        super(IngressRule, self).__init__(
+            spec_factory, ethertype, protocol, name)
+        self.reverse_class = EgressRule
+
+    @TrafficRuleBuilder.port_range.setter
+    def port_range(self, range_):
+        begin, end = self._port_range = range_
+        if begin:
+            self.ip_qualifier.destinationIpPort = self._port_range_spec(begin,
+                                                                        end)
+
+    @TrafficRuleBuilder.backward_port_range.setter
+    def backward_port_range(self, range_):
+        begin, end = self._backward_port_range = range_
+        if begin:
+            self.ip_qualifier.sourceIpPort = self._port_range_spec(begin, end)
+
+    @TrafficRuleBuilder.cidr.setter
+    def cidr(self, cidr):
+        self._cidr = cidr
+        if cidr:
+            self.ip_qualifier.sourceAddress = self._cidr_spec(cidr)
+
+
+class EgressRule(TrafficRuleBuilder):
+    direction = 'outgoingPackets'
+
+    def __init__(self, spec_factory, ethertype, protocol, name=None):
+        super(EgressRule, self).__init__(
+            spec_factory, ethertype, protocol, name)
+        self.reverse_class = IngressRule
+
+    @TrafficRuleBuilder.port_range.setter
+    def port_range(self, range_):
+        begin, end = self._port_range = range_
+        if begin:
+            self.ip_qualifier.destinationIpPort = self._port_range_spec(
+                begin, end)
+
+    @TrafficRuleBuilder.backward_port_range.setter
+    def backward_port_range(self, range_):
+        begin, end = self._backward_port_range = range_
+        if begin:
+            self.ip_qualifier.sourceIpPort = self._port_range_spec(
+                begin, end)
+
+    @TrafficRuleBuilder.cidr.setter
+    def cidr(self, cidr):
+        self._cidr = cidr
+        if cidr:
+            self.ip_qualifier.destinationAddress = self._cidr_spec(cidr)
+
+
+class DropAllRule(TrafficRuleBuilder):
+    action = 'ns0:DvsDropNetworkRuleAction'
